Move bot utilities to api.bots
