+++ docs/guides/bots/__init__.md
@@ -1,4 +1,4 @@
 A collection of guides for writing bots.
 
-[Write your first bot](simple.html)
-[Training bots](custom_battles.html)
+- [Write your first bot](simple.html)
+- [Training bots](custom_battles.html)
